requirements.txt
xnatpy
pandas
requests